+++ apps/web/src/features/auth/model/form/index.ts
@@ -1,5 +1,6 @@
 export { useEnrollmentVerificationForm } from './useEnrollmentVerificationForm';
 export { useEmailVerificationForm } from './useEmailVerificationForm';
+export { useFindPasswordForm } from './useFindPasswordForm';
 export { useOauthAdditionalInfoForm } from './useOauthAdditionalInfoForm';
 export { useSignUpAccountStep } from './useSignUpAccountStep';
 export { useSignUpEmailVerificationStep } from './useSignUpEmailVerificationStep';
